feat(sidebar): split explorer into per-feature FeatureItems
Replace DashboardSections with three focused components built on
FeatureItem: Connections, Favorites, and Recent. Each owns its own
selector and a shared SidebarNodeList, and exposes a view-all
IconButtonLink in its actions slot. The layout now renders the
self-contained ExplorerSidebar.
