+++ app/routes/layouts/protected.layout.tsx
@@ -10,10 +10,7 @@ import { authContext, authMiddleware } from "~/.server/auth/authMiddleware";
 import { toIdentity } from "~/.server/auth/getUserInfo";
 import { createLabel } from "~/.server/logging";
 import { PluginSlots } from "~/components/PluginSlots";
-import { ExplorerTab } from "~/components/Sidebar/Explorer/ExplorerTab";
-import { SIDEBAR_SEARCH_INPUT_ID } from "~/components/Sidebar/Explorer/SidebarSearchInput";
-import { Sidebar, SIDEBAR } from "~/components/Sidebar/Sidebar";
-import { useNavSidebarStore } from "~/components/Sidebar/sidebarStores";
+import { ExplorerSidebar } from "~/components/Sidebar/Explorer/ExplorerTab";
 import { useCredentialsKeepAlive } from "~/hooks/useCredentialsKeepAlive";
 import { useInitConnections } from "~/hooks/useInitConnections";
 import { loadFavorites } from "~/routes/favorites/favorites.loader";
@@ -76,15 +73,8 @@ export default function ProtectedLayout() {
     <div className="flex h-full flex-col">
       <PluginSlots name="app-banner" identity={identity} />
       <div className="relative flex flex-1 min-h-0">
-        <Sidebar
-          name={SIDEBAR.nav}
-          side="left"
-          store={useNavSidebarStore}
-          toggleShortcut="mod+b"
-          onOpen={() => document.getElementById(SIDEBAR_SEARCH_INPUT_ID)?.focus()}
-        >
-          <ExplorerTab />
-        </Sidebar>
+        <ExplorerSidebar />
+
         <div className="flex-1 overflow-x-hidden overflow-y-auto">
           <Outlet />
         </div>
